Clean path during file map build instead of during tree build
If a path is cleaned, then the lookup during tree build fails and this
causes unpredictable behavior in the SPK process. Cleaning paths as they
come in simplifies the code and makes it more robust.

While I was here, switched to "path" instead of "path/filepath" since
Docker/tar will always be using POSIX paths.

Fixes #12